requirements.txt
# requirements.txt - Install these packages

# Core bot functionality
python-telegram-bot>=20.0
openai>=1.50.0
python-dotenv==1.0.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Database
psycopg2-binary>=2.9.0

# Cache
redis>=4.6.0

# Queue
celery>=5.3.0

# Web Scraping
requests>=2.28.0
beautifulsoup4>=4.11.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
